requirements.txt
asgiref==3.2.10
azure-core==1.8.1
azure-iot-hub==2.2.2
azure-storage-blob==12.5.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1
Django==3.1.1
django-crispy-forms==1.9.2
django-jquery==3.1.0
djangorestframework==3.12.2
idna==2.10
isodate==0.6.0
msrest==0.6.19
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
uamqp==1.2.10
urllib3==1.25.10
